create verification.html
