+++ templates/verification.html
@@ -1,0 +1,21 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Verify Account</title>
+    <link rel="stylesheet" href="{{ url_for('static', filename='styles.css') }}">
+</head>
+<body>
+    <h2>Account Verification</h2>
+    <form method="POST" action="{{ url_for('verify') }}">
+        <label for="txtUSerEmail">Email:</label>
+        <input type="email" id="txtUSerEmail" name="txtUSerEmail" required><br><br>
+
+        <label for="txtVerificationCode">Verification Code:</label>
+        <input type="text" id="txtVerificationCode" name="txtVerificationCode" required><br><br>
+
+        <input type="submit" value="Verify">
+    </form>
+</body>
+</html>
